+++ CCHMapClusterController Example iOS/CCHMapClusterController Example iOS/ClusterAnnotationView.h
@@ -1,0 +1,15 @@
+//
+//  ClusterAnnotationView.h
+//  CCHMapClusterController Example iOS
+//
+//  Created by Hoefele, Claus(choefele) on 09.01.14.
+//  Copyright (c) 2014 Claus Höfele. All rights reserved.
+//
+
+#import <MapKit/MapKit.h>
+
+@interface ClusterAnnotationView : MKAnnotationView
+
+@property (assign, nonatomic) NSUInteger count;
+
+@end